fly_requirements.txt
flask
praw
python-telegram-bot<12.0
requests
gunicorn
APScheduler
trafilatura